Fix task assignment and extract NLD import service
